+++ src/addon_generator/ui/views/sampleprep_view.py
@@ -6,6 +6,7 @@ from PySide6.QtCore import QAbstractTableModel, QModelIndex, Qt
 from PySide6.QtWidgets import (
     QFormLayout,
     QHBoxLayout,
+    QHeaderView,
     QLabel,
     QLineEdit,
     QPushButton,
@@ -94,6 +95,8 @@ class SamplePrepView(QWidget):
         self.table.setModel(self.table_model)
         self.table.setSelectionBehavior(QTableView.SelectionBehavior.SelectRows)
         self.table.setSelectionMode(QTableView.SelectionMode.SingleSelection)
+        self.table.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeMode.ResizeToContents)
+        self.table.horizontalHeader().setStretchLastSection(True)
         splitter.addWidget(self.table)
 
         self.detail = DetailForm([FORM_LABELS[name] for name in FIELD_ORDER], self)